+++ src/Launcher.java
@@ -1,7 +1,7 @@
 import support.SunHttpServer;
 import yose.Routes;
-import yose.http.Router;
 import yose.http.Server;
+import yose.http.routing.Router;
 
 import java.io.IOException;
 
@@ -10,7 +10,7 @@ public class Launcher {
     public static void main(String[] args) throws IOException {
         int port = Integer.parseInt( args[ 0 ] );
         Server server = new SunHttpServer( port );
-        server.useRouter( Router.routing( new Routes() ) );
+        server.useRouter(Router.routing(new Routes()));
         server.start();
     }
 }
