+++ src/support/SunHttpServer.java
@@ -2,7 +2,10 @@ package support;
 
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpServer;
-import yose.http.*;
+import yose.http.HttpRequest;
+import yose.http.HttpResponse;
+import yose.http.Server;
+import yose.http.routing.Router;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
