--- src/yose/http/HttpRequestMatcher.java
@@ -1,7 +1,0 @@
-package yose.http;
-
-public interface HttpRequestMatcher {
-
-    boolean matches(HttpRequest request);
-}
-
