+++ src/yose/http/Server.java
@@ -1,5 +1,7 @@
 package yose.http;
 
+import yose.http.routing.Router;
+
 public interface Server {
 
     void useRouter(Router router);
