+++ test/support/YoseChallenge.java
@@ -3,7 +3,7 @@ package support;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import yose.Routes;
-import yose.http.Router;
+import yose.http.routing.Router;
 
 public class YoseChallenge {
 
@@ -12,7 +12,7 @@ public class YoseChallenge {
     @BeforeClass
     public static void startServer() throws Exception {
         server = new SunHttpServer( 8000 );
-        server.useRouter( Router.routing( new Routes() ) );
+        server.useRouter(Router.routing(new Routes()));
         server.start();
     }
